Refactor(resources/js/components/) use new folder structure (feature folders)
