+++ resources/js/app.js
@@ -24,4 +24,8 @@ addHashActiveTab();
  * or customize the JavaScript scaffolding to fit your unique needs.
  */
 
-require('./components/index');
+require('./components/codeEditor/index');
+
+
+// require('./components/editor/index');
+// require('./components/somethingElse'); // если появится что-то новое
