--- resources/js/components/Editor/ControlBox.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import { Button } from 'react-bootstrap';
-
-const ControlBox = () => {
-
-  return (
-    <div className='d-flex py-2 w-100 justify-content-end card-footer py-3'>
-      <Button variant="primary">Запустить</Button>
-    </div>
-  );
-};
-
-export default ControlBox;
