+++ resources/js/components/codeEditor/app/init.jsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import App from './App.jsx'
+
+export default async () => <App />;
