+++ resources/js/components/codeEditor/features/editor/style.css
